backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pandas==2.1.4
numpy==1.26.3
pyyaml==6.0.1
networkx==3.2.1
celery==5.3.6
croniter==2.0.1
redis==5.0.1
slowapi==0.1.9
python-multipart==0.0.6
python-dotenv==1.0.0
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==5.0.0
httpx==0.26.0
factory-boy==3.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
flower==2.0.1
prometheus-fastapi-instrumentator==6.1.0
sentry-sdk[fastapi]==1.39.1
orjson==3.10.15
boto3==1.34.0
